Updates to visualization
